Add a device tree file for the CETiBOX X3 and activate CAN-FD controller
The device tree is based on r8a7796-m3ulcb.dtb, which is the board
currently used for development.

The classic CAN controller is currently disabled in favor of CAN-FD,
as there seem to be no way to use both in parallel and CAN-FD is more
flexible. Subject to change, though.

CBX3DEV-4