Add enable/suspend GPIOs to the CAN controller device tree
Taken from the patches in CogentEmbedded repository for the "ADAS board"
support. In particular, the patch

0033-arm64-dts-r8a7795-es1-h3ulcb-had-add-ADAS-board.patch

inpired the choice of the GPIO pins: the rumors are that Renesas used
the H3ULCB.HAD board (aka "ADAS board") to verify CAN functionality in
H3 starter kit. Which is yet another different board, but might be
similar enough to the M3 board.

Note: it does not work: the driver has a GPIO conflict when enabling can0 and can1
(cherry picked from commit 1413492afea67e08d7243ae37c9071c575646920)