Use the CAN-FD controller instead of classic CAN
The internal classic CAN controller does not support reception
of base and extended CAN frames at the same time and the SoC documentation
seems to be confused.

With the driver modified to support suspend/enable pins per channel
and the CAN-FD controller switched to classic CAN mode this controller
seems to be a better choice.

CBX3DEV-4