Activated PWM2 for hard-resetting the SJA1105 switch.
Unfortunately the corresponding pin cannot be configured as a GPIO, so
we will use PWM2 and generate a PWM signal with positive polarity, long
period and short duty cycle, and activate it for a short time during
bootup. This does the job - we tested it.

(cherry picked from commit 20f7168ef15268f6fe8e27c57fef09afa0bb71ef)