[ethernet: ravb] Disable NC Rx queue for SJA1105 meta frames
When using the SJA1105 switch with gPTP, the Rx timestamps are sent as
metaframes directly following the received frame. We have to disable the
NC queue and receive gPTP in the BE queue, because otherwise the gPTP
frame and the following metaframe end up in different queues and the
mapping between gPTP and the associated metaframe is lost.

(cherry picked from commit 1886f45502a07cde7a5d48b949f1c565e873873d)