Completely disable OTG on channel 0 of EHCI controller
The channel is used exclusively for the INIC device. As the automatic
selection by ID pin is broken on the gateway board, hard-code the
channel mode in the driver for a while.

The ID line of the channel is left unconnected, but should be grounded
to put the channel in host mode.

CBX3IST-79